--- a/english/main.go
+++ b/english/main.go
@@ -31,7 +31,7 @@ func NewSentenceTokenizer(s *sentences.Storage) (*sentences.DefaultSentenceToken
 	}
 
 	// supervisor abbreviations
-	abbrevs := []string{"sgt", "gov", "no"}
+	abbrevs := []string{"sgt", "gov", "no", "mt"}
 	for _, abbr := range abbrevs {
 		training.AbbrevTypes.Add(abbr)
 	}