Hi,

I did a pass over the Golden Rules and I was able to get the failing cases (excluding the list-related tests) down to 4:

```
--- FAIL: TestGoldenRules (0.00s)
        golden_rules_test.go:11: 14. Multi-period abbreviations at the end of a sentence
        golden_rules_test.go:12: Actual: [<Sentence [0:33]>]
        golden_rules_test.go:13: Actual: 1, Expected: 2
        golden_rules_test.go:14: ===
        golden_rules_test.go:11: 15. U.S. as sentence boundary
        golden_rules_test.go:12: Actual: [<Sentence [0:33]>]
        golden_rules_test.go:13: Actual: 1, Expected: 2
        golden_rules_test.go:14: ===
        golden_rules_test.go:11: 18. A.M. / P.M. as non sentence boundary and sentence boundary
        golden_rules_test.go:12: Actual: [<Sentence [0:37]> <Sentence [37:98]>]
        golden_rules_test.go:13: Actual: 2, Expected: 3
        golden_rules_test.go:14: ===
        golden_rules_test.go:20: 43. Geo Coordinates
        golden_rules_test.go:21: Actual: [You can find it at N°.] Expected: [You can find it at N°. 1026.253.553.]
        golden_rules_test.go:22: ===
```

My changes are all local to the `english` package and all other tests are still passing. Here's the performance impact on the English test files:

```
name              old time/op  new time/op  delta
EnglishPackage-4  8.70ms ± 1%  9.42ms ± 1%  +8.31%  (p=0.008 n=5+5)
```

I'd like to hear your thoughts on my progress so far.

Thanks for your work on this useful package!