@neurosnap 2017-04-15
Hey!  Thanks for taking the time to try and improve the library, I really appreciate it!

At first glance this looks really great!  I'm going to take some more time soon to review the code, but using the English package was exactly where I was expecting this work to go, so props to you.

Again, really great work!