@jdkato 2017-04-16
Thanks for the review!  I cleaned up the tests and removed the benchmarks.