@neurosnap 2017-05-18
Hey @jdkato 

Sorry I haven't looked at this PR in awhile.  My main apprehension is ensuring that we didn't inadvertantly break any of our other examples.

The core of my tests were comparing my library against nltk's library even knowing that nltk has mistakes.

I'd like to run this PR against the examples in the main library and see if there are any regressions that we have noticed from these changes.

What are your thoughts on this?