@jdkato 2017-05-18
Hey,

After using this library extensively for a while now, I've come to realize that it's not a great fit for my needs&mdash;so, I'd be fine with just closing this out. If you still want to move forward, though, I'd be glad to help wherever I can.