@neurosnap 2017-05-18
Ok, I'm going to spend some time soon experimenting with this PR.

What are you using instead of this library, if you don't mind me asking?