@jdkato 2017-05-19
I've been using the Pragmatic Segmenter itself for other work, and I think its rule-based approach (i.e., no training data) fits my particular Go needs (a multilingual, standalone CLI tool) better too. So, I'm going to start working on a native Go port, which I hope to finish in a week or so.